feat: 로그인 없이 profile 접근시 로그인 페이지로 리다이렉트 && style: 초기화면 로그인 여하에 따른 렌더링 스타일링
