+++ index.html
@@ -5,7 +5,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>항해플러스 SNS</title>
     <script src="https://cdn.tailwindcss.com"></script>
-    <script src="./src/router.js" type="module"></script>
+    <link rel="stylesheet" href="/src/style/index.css" />
+
+
   </head>
   <body>
     <nav>
