+++ src/style/index.css
@@ -1,0 +1,3 @@
+.display-hidden{
+    display: none ;
+}
